+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/CameraInfoProxyApi.java
@@ -61,4 +61,9 @@ class CameraInfoProxyApi extends PigeonApiCameraInfo {
     return new LiveDataProxyApi.LiveDataWrapper(
         pigeonInstance.getZoomState(), LiveDataSupportedType.ZOOM_STATE);
   }
+
+  @Override
+  public boolean hasFlashUnit(CameraInfo pigeonInstance) {
+    return pigeonInstance.hasFlashUnit();
+  }
 }