+++ packages/camera/camera_android_camerax/android/src/test/java/io/flutter/plugins/camerax/CameraInfoTest.java
@@ -99,4 +99,15 @@ public class CameraInfoTest {
 
     assertEquals(value, api.getZoomState(instance).getLiveData());
   }
+
+  @Test
+  public void hasFlashUnit_makesCallToRetrieveHasFlashUnit() {
+    final PigeonApiCameraInfo api = new TestProxyApiRegistrar().getPigeonApiCameraInfo();
+
+    final CameraInfo instance = mock(CameraInfo.class);
+    final boolean value = true;
+    when(instance.hasFlashUnit()).thenReturn(value);
+
+    assertEquals(value, api.hasFlashUnit(instance));
+  }
 }